Add support for collections of ints and strings to pgonet.
Global state management now supports collections of ints and strings.

Caveats: when setting a collection, the entire collection needs to be
passed and thus uploaded to etcd. The same happens when retrieving the
values of collection types. This approach is simple and naive, and will
certainly not be optimal in terms of performance.

Global state management needs to be completely re-thought moving forward
in this project.