Allow WaitPeers() to be called multiple times.
With a relatively simple change in the implementation of
`StateServer.WaitPeers()`,  it is now possible to call the function
whenever a barrier across all nodes in the system is necessary.

This comes in handy because we want to have a synchronization barrier at
the end of the execution of a finite PlusCal algorithm, to make sure
that a process will only terminate once all processes are done. If a
node is finished running its PlusCal algorithm, it remains a part of the
system and is able to own global state if necessary.