Make ArchetypeResources comparable.
This extends the ArchetypeResource interface to include a Less function, which
is supposed to account for the fact that global variables are sensitive to the
order in which they are acquired (using the wrong order could lead to
deadlocks).

The `Less` function enables ArchetypeResources to be sorted by using Go's
sort.Sort function. The current implementation only does something meaningful
for archetype resources of type GlobalVariable. When comparing two global
variables, the Less function will perform lexicographical comparison. For all
other cases, `false` is returned, indicating that the two resources being
compared should not be swapped (are not comparable).

This adds some runtime cost to the process of acquiring/releasing resources,
but, in general, the number of parameters an archetype is going to receive is
going to be small, and the runtime cost should be negligible (although actual
numbers were not collected.) However, this change does completely free the
user/caller from having to explicitly specify an order of archetype resources
when invoking functions generated by PGo.