Make Pcal / Mpcal parsers run in less than exponential time
Due to some omissions in the PlusCal (and thus MPCal) parser, there was redundant nondeterminism in the way
the parser would read blocks of statements. As a result, it would get stuck reparsing the same pieces of code
exponentially more times the longer and more complex a block was as the ambiguities were compounded over sequences
of statements.

This is exactly what the `cut()` operator was designed to prevent - this commit adds some calls to `cut()` that avoid
nodeterminism when the first parsing result is guaranteed to be the correct one.

Specifically, in these cases:
- TLA+ expressions: these are always followed by something that is unambiguously not part of a TLA+ expression
  (in fact, they are already almost always `cut()`, it seems I forgot that when writing the PlusCal parser).
  In general, watch out for TLA+ expressions. The way their grammar is written can leave a lot of ambiguity, so
  `cut()` them unless you know why not to (there is one case of this, the `a_<<b,c>>(d)` construct)
- the beginning of a `{ ... }` block: a `{` is unambiguous whenever these might occur in PlusCal (not to be confused
  with other notations in TLA+, which will never occur in the same context as the PlusCal `{`)
- some miscellaneous locations where repetition is involved. repetitions generate ambiguity as a feature
  (look at how they are implemented), but it is not needed here.