Improved TLA+ parse phase speed via hack
I tried a few things to make the TLA+ parser faster, but the more I try the more it occurs to me that the design itself is slow.
There's no obvious exponential backtracking when I trace parser execution, so nothing that can be fixed without a much more sophisticated
tweak to parser control flow (which would affect the most complex and bug-prone part of the parser: the expression parser).

As such, for the sake of getting results now I disabled parsing the translated TLA+ code. This nets a significant speedup, since
most of the TLA+ PGo had to parse was in there. This broke one test which relied on capturing units during / after the translation,
which I commented out with a note as to why.

I hope to return one day with a proper fix for this, which may be an actually fast parser based on my thesis work :)