LE and EQ codegen for maps mostly complete
In principle, the structure is there now. I re-used the slice
comparators, adapting maps to them by extracting slices of keys and
values. Neither pretty nor particularly efficient, but general and
should get the job done. This codegen system isn't well suited to
fancy optimisations anyway (TODO).

The only problem is that my test case is having a map value type
inferred as `interface{}`, which is not in any way comparable...
tried to debug the typing judgement that lead to this but I'm out
of ideas for the evening.

Commented out test cases are in TypeSolverTest. They pass - why do
they do the right thing when codegen doesn't? Something I'm missing?